requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
cryptography==42.0.5
pyjwt==2.8.0
httpx==0.27.0
pytest==8.1.1
pytest-cov==5.0.0
requests==2.31.0
ruff==0.3.2
